Pin droidmedia to 0.20211101.0
In SailfishOS, gst-droid is built against droidmedia-devel
which is provided by the adaptation-common repository.

In the current (4.4) release of Sailfish, droidmedia is expected
to be 0.20211101.0.  Because on Halium, droidmedia is provided by
the halium rootfs, the version must match the version in the current
release because changes in the headers used to build gst-droid may
result in crashes if there is a mismatch.

I would like to keep the version used in Halium in line with the
current released SailfishOS version to minimise problems.

I dont think there are currently any other users of droidmedia and
gst-droid so this should have minimal impacts.